requirements.txt
# Core frontend/backend
streamlit==1.35.0
fastapi==0.110.1
uvicorn==0.29.0
requests==2.31.0
pandas==2.2.2

# LangChain core
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-huggingface==0.0.6

# Embeddings & vector store
sentence-transformers==2.6.1
faiss-cpu==1.8.0

# PDF/text handling
pytesseract==0.3.10
pdfminer.six==20221105

# Utility
python-dotenv==1.0.1
